delete button work! in profile-page and in feed

+++ src/app/components/feed/feed.component.ts
@@ -47,7 +47,9 @@ export class FeedComponent implements OnInit {
   }
 
   deleteMessage(post: any){
-    console.log(post);
-    this.msgService.deletePost(post.id)
+    if(confirm(this.user.displayName+", Are you sure you want to delete this post?")) {
+      console.log(post);
+      this.msgService.deletePost(post.id)
+    }
   }
 }
